Joe Armstrong
Why OO Sucks http://www.cs.otago.ac.nz/staffpriv/ok/Joe-Hates-OO.htm
Joe Armstrong said “The problem with object-oriented languages is they've got all this implicit environment they carry around with them. You wanted a banana but what you got was a gorilla holding the banana and the entire jungle.” How true is this? https://www.quora.com/Joe-Armstrong-said-The-problem-with-object-oriented-languages-is-theyve-got-all-this-implicit-environment-they-carry-around-with-them-You-wanted-a-banana-but-what-you-got-was-a-gorilla-holding-the-banana-and-the