Vertically scaling Kafka consumers https://www.jesseyates.com/2019/12/04/vertically-scaling-kafka-consumers.html
Vertically scaling Kafka consumers https://www.jesseyates.com/2019/12/04/vertically-scaling-kafka-consumers.html
A look at the inner workings of the Kafka consumers, with some real world recommendations for deploying them when there's high latency in talking to the Kafka cluster and/or a large number of partitions. There are tips on important metrics to monitor, configurations, garbage collector settings, and changing the partition.class to improve unbalanced consumers.