Running and Deploying Elasticsearch on Kubernetes https://sematext.com/blog/kubernetes-elasticsearch
Running and Deploying Elasticsearch on Kubernetes https://sematext.com/blog/kubernetes-elasticsearch
An overview of Kubernetes and how to deploy Elasticsearch on Kubernetes. This is a great introduction to many of the core concepts of Kubernetes (e.g. Deployment, Pod), including those that are important for running a stateful service (e.g. StatefulSet, PersistentVolumeClaim). It also shows how to configure accounts for ES using Kubernetes RBAC and use the helm package manager for deploying to Kubernetes.