bc, for
`echo "100 / 3" | bc -l` or `bc -l <<< '100 / 3'` How do I use floating-point division in bash? https://stackoverflow.com/questions/12722095/how-do-i-use-floating-point-division-in-bash
`echo $1 + $2 | bc` Addition of two floating point numbers using shell script https://unix.stackexchange.com/questions/360324/addition-of-two-floating-point-numbers-using-shell-script

Linux bash. for loop and function, for adding numbers https://stackoverflow.com/questions/22460266/linux-bash-for-loop-and-function-for-adding-numbers/22460518
`for i in $(seq -f "$02g" 1 20); do echo $i; done` How to zero pad a sequence of integers in bash so that all have the same width? https://stackoverflow.com/questions/8789729/how-to-zero-pad-a-sequence-of-integers-in-bash-so-that-all-have-the-same-width