VSCode, WSL
Visual Studio Code Docker extension 1.0 – Better than Ever! https://devblogs.microsoft.com/visualstudio/visual-studio-code-docker-extension-1-0-better-than-ever

How to Create a List of Your Installed Programs on Windows https://www.howtogeek.com/165293/how-to-get-a-list-of-software-installed-on-your-pc-with-a-single-command

Docker Desktop for Windows Home is here! https://www.docker.com/blog/docker-desktop-for-windows-home-is-here
WSL2 will be generally available in Windows 10, version 2004 https://devblogs.microsoft.com/commandline/wsl2-will-be-generally-available-in-windows-10-version-2004