Aliases, Simple and Efficient, gops, slices, Goroutine
Go: Aliases, Simple and Efficient https://medium.com/a-journey-with-go/go-aliases-simple-and-efficient-8506d93b079e

Go: How Does gops Interact with the Runtime? https://medium.com/a-journey-with-go/go-how-does-gops-interact-with-the-runtime-778d7f9d7c18

Golang tips: why pointers to slices are useful and how ignoring them can lead to tricky bugs https://medium.com/swlh/golang-tips-why-pointers-to-slices-are-useful-and-how-ignoring-them-can-lead-to-tricky-bugs-cac90f72e77b

Go: How Does a Goroutine Start and Exit? https://medium.com/a-journey-with-go/go-how-does-a-goroutine-start-and-exit-2b3303890452