Arrow Function, Clean Code, TypeScript 4.0
Do you really know when to use an Arrow Function in JavaScript? https://medium.com/javascript-in-plain-english/do-you-really-know-when-to-an-arrow-function-in-javascript-fdcbcdae7297

Who Else Wants to Write Clean JavaScript Code? - JavaScript In Plain English - Medium https://medium.com/javascript-in-plain-english/who-else-wants-to-write-clean-javascript-code-ff4f7896e6bd

What Can We Expect in TypeScript 4.0? - Better Programming - Medium https://medium.com/better-programming/what-can-we-expect-in-typescript-4-0-b5be85021b90