Segmentation fault: 11, generate a random number, difference between the new operator and malloc()
Joe Zbiciak's answer to In C++, why do I get a 'Segmentation fault: 11' error when I attempt to access an array with more than 525 elements? - Quora https://www.quora.com/In-C-why-do-I-get-a-Segmentation-fault-11-error-when-I-attempt-to-access-an-array-with-more-than-525-elements/answer/Joe-Zbiciak

David Vandevoorde's answer to How do you generate a random number between 1 and 6 in C++? - Quora https://www.quora.com/How-do-you-generate-a-random-number-between-1-and-6-in-C/answer/David-Vandevoorde

Joe Zbiciak's answer to In C++, what is the difference between the new operator and malloc() function (In general and especially for objects allocation)? - Quora https://www.quora.com/In-C-what-is-the-difference-between-the-new-operator-and-malloc-function-In-general-and-especially-for-objects-allocation/answer/Joe-Zbiciak