Benefits of Continuous Delivery
Benefits of Continuous Delivery https://www.facebook.com/groups/474127499987742?multi_permalinks=1148371552563330
The main objective of #continuousdelivery (CD) is cycle time reduction.
It is faster because it is based on automated tasks; waiting time is eliminated.
It is cheaper because automated tasks allow errors to be detected immediately, avoiding the accumulation of defects.
It is better because automation provides predictable, standardized, reliable and repeatable results.